microservices/csam-scanner/requirements.txt
fastapi
numpy
opencv-python-headless
pdqhash
vpdq
mysql-connector-python
